Code cleanup Bris_T1D Dataloader (#228)
* Got rid of old debugging notebook.

* Transfered bris_t1d data cleaning functions to the data_cleaner.py module.

* Made recommneded changes from co-pilot code review, fixed ruff formatter.
